+++ front/gatsby/package.json
@@ -35,7 +35,6 @@
   },
   "scripts": {
     "build": "gatsby build",
-    "netlify": "gatsby build",
     "start": "gatsby develop",
     "start:prod": "yarn build && yarn serve",
     "serve": "gatsby serve",
